+++ query/transaction.go
@@ -43,10 +43,11 @@ type (
 	}
 	TransactionControl  = tx.Control
 	TransactionSettings = tx.Settings
+	TransactionOption   = tx.Option
 )
 
 // BeginTx returns selector transaction control option
-func BeginTx(opts ...tx.Option) tx.ControlOption {
+func BeginTx(opts ...TransactionOption) tx.ControlOption {
 	return tx.BeginTx(opts...)
 }
 
@@ -111,19 +112,19 @@ func TxSettings(opts ...tx.Option) TransactionSettings {
 	return opts
 }
 
-func WithDefaultTxMode() tx.Option {
+func WithDefaultTxMode() TransactionOption {
 	return tx.WithDefaultTxMode()
 }
 
-func WithSerializableReadWrite() tx.Option {
+func WithSerializableReadWrite() TransactionOption {
 	return tx.WithSerializableReadWrite()
 }
 
-func WithSnapshotReadOnly() tx.Option {
+func WithSnapshotReadOnly() TransactionOption {
 	return tx.WithSnapshotReadOnly()
 }
 
-func WithStaleReadOnly() tx.Option {
+func WithStaleReadOnly() TransactionOption {
 	return tx.WithStaleReadOnly()
 }
 
@@ -131,6 +132,6 @@ func WithInconsistentReads() tx.OnlineReadOnlyOption {
 	return tx.WithInconsistentReads()
 }
 
-func WithOnlineReadOnly(opts ...tx.OnlineReadOnlyOption) tx.Option {
+func WithOnlineReadOnly(opts ...tx.OnlineReadOnlyOption) TransactionOption {
 	return tx.WithOnlineReadOnly(opts...)
 }